requirements.txt
python-telegram-bot==21.3
requests==2.31.0
python-dotenv==1.0.0
aiohttp==3.9.5
pydantic==2.7.1
loguru==0.7.2
pytz==2024.1